Added the basic interfaces for the DaemonManager and ManagedDaemon.
